Fix the is_so_mark_functional test in Linux containers
It's not sufficient to check for root privileges.  In a container, the
euid may be root even though the user lacks some capabilities.  Replace
this test's root check with a check for the CAP_NET_ADMIN capability
instead.
