+++ Cargo.toml
@@ -30,6 +30,9 @@ lazy_static = "1.2"
 rand = "0.6"
 tempfile = "3.0.5"
 
+[target.'cfg(any(target_os = "android", target_os = "linux"))'.dev-dependencies]
+caps = "0.3.1"
+
 [target.'cfg(target_os = "freebsd")'.dev-dependencies]
 sysctl = "0.1"
